Port of D787 to od4.6( Made Angle Range parameters int instead of float )
Summary: same as D787

Test Plan: same as D787

Reviewers: raman

Reviewed By: raman

CC: helene

Differential Revision: https://phabricator.dgbes.com:8443/D789

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@29814 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/PreStackProcessing/prestackanglemute.h
@@ -39,7 +39,7 @@ public:
 				AngleCompParams();
 					
     float 			mutecutoff_;
-    Interval<float>		anglerange_;
+    Interval<int>		anglerange_;
     MultiID			velvolmid_;   
     IOPar			raypar_;
     IOPar			smoothingpar_;
